fix(srtla_send): drop the size-based keyframe heuristic, trust the hint
The critical-routing override fired on `heuristic_keyframe || window_critical`,
where heuristic_keyframe was a guess from runs of 1316-byte packets. srtla_send
only sees opaque SRT payloads, so it cannot distinguish a keyframe from any
other run of max-MTU packets; at sustained high bitrate the heuristic
false-positived and pinned ordinary traffic to a single link, defeating
bonding.

Keyframe detection is authoritative upstream: the encoder front-end (gst-app)
parses NAL units and opens the critical window over the priority sidecar. Drop
the heuristic entirely and fire the override only on that window. The lone
remaining helper (select_best_quality_idx) moves into priority.rs beside the
CriticalWindow it serves; src/sender/keyframe.rs is deleted.
